move to a trait
